+++ revdep/library.noindex/WeibullR.shiny/old/WeibullR.shiny/app/dependencies.R
@@ -1,0 +1,4 @@
+require(WeibullR)
+require(WeibullR.plotly)
+require(shiny)
+require(shinydashboard)