+++ revdep/library.noindex/WeibullR.shiny/old/WeibullR.shiny/app/global.R
@@ -1,0 +1,4 @@
+library(WeibullR)
+library(WeibullR.plotly)
+library(shiny)
+library(shinydashboard)