+++ news.md
@@ -1,9 +1,15 @@
-# WeibullR.shiny 0.2.1
+# WeibullR.shiny v0.3
+
+* Now with support for Reliability Growth Analysis including Crow-AMSAA, Piecewise Weibull NHPP, Change Point Detection, and Duane models.
+* New tutorials on Time-to-Failure and Reliability Growth Analysis.
+* Other minor improvements and bug fixes.
+
+# WeibullR.shiny v0.2.1
 
 ## Minor improvements and bug fixes
 * Updated contact info and citation.
 
-# WeibullR.shiny 0.2
+# WeibullR.shiny v0.2
 
 ## Minor improvements and bug fixes
 * Now with a lognormal plotting canvas.